02_fashion_mnist/requirements.txt
ipython==5.8.0
Keras-Applications==1.0.5
Keras-Preprocessing==1.0.3
matplotlib==3.0.0
numpy==1.15.2
Pillow==5.3.0
prompt-toolkit==1.0.15
ptipython==1.0.0
ptpython==0.41
pudb==2018.1
tensorflow==1.11.0
